tests: Ignore ENOENT errors when deleting temporary files
When cleaning up after a test has run, ignore ENOENT errors when
deleting files, because they basically indicate success.

I suspect this is happening because the gpg-agent process is being
killed at the same time as we are listing and deleting files, and it
removes its own socket file on exit, after we’ve listed temporary files
to delete, but before we’ve actually deleted it.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T18308