apply: Add a mechanism to atomically update flatpaks with OS upgrades
The updater now looks in /usr/share/eos-application-tools/flatpak-autoinstall.d/*
to for an action list of flatpaks to either install or remove from the system
before deploying a certain commit. It works by keeping a parallel copy of the
file in /var/lib/eos-application-tools/flatpak-autoinstall.d/* and taking
the difference between the two to determine what actions need to be applied
before the commit is deployed.

The format of the file is as follows

   action remote:ref

where action is one of either "install" or "remove" and remote:ref is a valid
flatpak refspec. The branch and architecture part of the refspec must be
specified, but will be ignored.

Note that it assumes that the file is only ever append only - if you need
to remove a flatpak in a later update, you should add a new line with
a "remove" action. It is completely valid to have a file that specifies
that a flatpak should be removed, then installed, then removed again -
the mechansim will squash down changes such that only a minimal amount
of work is done.

Before a system update is applied, the updater will now look inside
the commit to determine which flatpaks to install or remove. In the
case where flatpaks need to be installed, it will pull the relevant
ostree refs into the flatpak repository and only on success continue
with deployment of the system update. Then, once the deployment occurrs,
a second script will run on boot which examines the contents of
both /var/lib/eos-application-tools/flatpak-autoinstall.d/* and
/usr/share/eos-applicationt-tools/flatpak-autoinstall.d/* and
will either deploy flatpaks to remove them appropriately. Thus,
an internet connection is only required up to the point when
the system update is deployed, but is not required on the next
reboot when flatpaks are actually installed or removed.

https://phabricator.endlessm.com/T16682