libeos-updater-util: Disable some compile warnings for SipHash code
The imported SipHash code triggers a few compiler warnings which causes
build failures with -Werror. Avoid that by explicitly disabling them
using a pragma (should work on gcc and clang) for the SipHash code only.
This keeps the SipHash code unmodified for easier verification and
diffability to upstream.

Also disable -Wpragmas so that GCC versions older than 7 don’t complain
about the -Wimplicit-fallthrough warning (which was added in GCC 7).

Signed-off-by: Philip Withnall <withnall@endlessm.com>