Remove Volume Path from eos-updater.conf
Since eos-updater will use ostree_repo_find_remotes_async() instead of
searching for LAN and USB remotes itself, the configuration key
specifying a USB volume is no longer needed.

This commit deletes the code using the key which breaks it, but it's the
code that will be deleted in a following commit anyway.

https://phabricator.endlessm.com/T19293