eos-updater-prepare-volume: Rewrite in Python and use ostree-create-usb
This is a complete rewrite of the eos-updater-prepare-volume utility,
although it maintains command line backwards compatibility with the old
version.

The new version is written in Python, and is a fairly simple wrapper
around `ostree create-usb` (requires OSTree 2017.12). It adds support
for including flatpaks in the repository on the USB drive, resolving
various details about them from the provided collection–refs.

All the code relating to the old version of the utility has been
deleted. This enables us to delete the libeos-updater-0 library and its
associated Debian rules. As far as I’m aware, nothing outside
eos-updater ever used it.

Signed-off-by: Philip Withnall <withnall@endlessm.com>
https://phabricator.endlessm.com/T19366