src: Fix paranoid code for invalid calls to eos_updater_set_error()
eos_updater_set_error() had some paranoid code which tried to handle the
cases where it’s called incorrectly (with error == NULL) by using
default values for the various GError fields. However, this ignored the
g_dbus_error_encode_gerror() call, which blows up when passed an invalid
or NULL GError, with assertopm failures.

Fix that by overwriting a NULL GError with a consistent, newly allocated
error, so that we can guarantee all the fields are set properly.

This retains the paranoia code on the assumption that it’s better to
report a default error to the user than it is to hit a precondition
failure and skip reporting any error.

This was found as a symptom of the error corruption bug fixed in the
following commit.

Signed-off-by: Philip Withnall <withnall@endlessm.com>