tests: Fix OSTree-based tests running on an ARM chroot
When an OstreeRepo is opened, it looks at the boot ID provided by the
kernel to find some uniqueness. In an ARM chroot, that doesn’t
necessarily exist, causing tests to fail.

In the case of the libeos-updater-util/tests/ostree test, we can set the
boot ID using OSTREE_BOOTID in the environment, and the test now passes.

For the tests/test-update-* tests, we could do that, but the tests
subsequently fail with problems with xattrs (qemu-user doesn’t implement
the necessary syscalls) — so use the existence of the boot ID as a proxy
for ‘running in a chroot’ and skip those tests if so.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T19388