src: Explicitly cast a GPOINTER_TO_INT() return value to an enum type
When compiling with -Wconversion with clang (not with gcc), this line
gave an error, presumably because clang treats the UpdateStep enum as
unsigned, whereas gcc treats it as signed. The C specification leaves
this as implementation-specific (thanks, C).

Fix that by making the cast from signed int to whatever-the-enum-is
explicit, rather than implicit. This introduces no functional changes,
but does shut clang up.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T19388