src: Drop outdated LAN and USB poll code
Since we now use OstreeRepoFinder to find LAN and USB update sources, we
don’t need the custom code which eos-updater previously had. Drop it.

There is some integration in eos-updater-poll-volume-dbus.c which needs
reworking, as we should keep the ability to poll a specified repository
URI, but that will be reworked in a following commit.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T19293