src: Support chained endoflife-rebase redirects
Previously, the code only supported following a single endoflife-rebase
key in the polled commit metadata. While unlikely, we should technically
support following a chain of them to get to the first commit which
doesn’t have an endoflife-rebase key.

Split the code to parse commit metadata out from that which fetches the
commit metadata, and implement a lookup loop.

Signed-off-by: Philip Withnall <withnall@endlessm.com>