src: Add OstreeRepoFinder code paths in poll and fetch
Use OstreeRepoFinder as the preferred way to find and fetch updates from
main, LAN and USB sources. The old code still exists, but will be
dropped (for LAN and USB, but not for main) in a following commit.

The critical part of this approach is that the OstreeRepoFinderResult
array is passed from the poll to the fetch stage via the EosUpdaterData
struct.

Signed-off-by: Philip Withnall <withnall@endlessm.com>