src: Add support for override URIs when polling
Use the new OstreeRepoFinderOverride implementation in libostree to
support passing override URIs into the OstreeRepoFinder machinery.

This will be used in a following commit to rework the tests, so we can
avoid having to simulate USB drives or a LAN. It is also used by the
PollVolume call, which is the public dynamic interface for polling from
a given URI. The code for that has moved into eos-updater-poll.c, since
keeping it in a separate file wasn’t conducive to sharing code.

The config file option for setting the override URIs may be documented
in future. For the moment, it remains semi-private, and may be removed
or changed in future.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T19293