tests: Rework LAN and USB test harnesses to use override URIs
Rather than hooking into the now-removed emulation code for LAN and USB
sources, just set an override URI when doing LAN or USB tests. This
avoids any OstreeRepoFinder{Avahi,Mount} code, and instead uses
OstreeRepoFinderOverride in the poll code.

That means we’re not doing integration testing of the DNS-SD side of
finding LAN peers; and similarly not testing the volume-finding side of
finding USB sources — but we weren’t really doing that before, and
realistically can’t do that without emulating something at some level,
which would be unrepresentative. The eventual plan is to test this part
of the LAN and USB code by running integration tests on complete Endless
OS images.

What these tests do achieve, however, is to test the server side of LAN
and USB sources — eos-update-server and eos-updater-prepare-volume. It
also tests everything after the poll stage, including fetching and
everything to do with commit signing and verification.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T19293