src: Split out metrics upload code for successful polls
This will be reused in another code path in an upcoming commit. This
commit also fixes a latent bug where the old ref name would be reported
in the metric, rather than the new ref name. The ref name can change
when we do an update, as a result of endoflife-rebase metadata in the
commits.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T19293