eos-updater-avahi: Fix last-modified timestamp lookup from summary file
The code was trying to look up the last-modified timestamp from the
top-level variant in the summary file, rather than the metadata a{sv}
which is one of its children. It was also not validating that the
GVariant was in normal form before continuing, which we now have to do
due to summaries being unsigned.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T19293