flatpak: Use collection-id to specify flatpak source, not remote-name
This is a little bit tricky to implement, considering that flatpak
doesn't have support for collection-ids right now, but we want
to future-proof the file format against the inevitable upcoming
change. Because of the lack of support in flatpak, we have to
"simulate" collection-refs by looking up which remotes
support a collection ref and then using the first one that comes
up.

It gets especially hairier when we want to actually deploy the
flatpaks since we have no guarantee of network access at that
point. The best we can do is look up which remote the collection-id
probably came from and use that when installation, as opposed to
actually downloading the summary file from the remote.