flatpak: Slightly rework the way that flatpak ref actions are read
Instead of an artificial distinction between the base directory
and override directories, have a single function that takes
a GStrv of directories in priority order and returns a
GHashTable of remote ref actions. Then we can wrap that function
in another one which appends the lowest priority base directory
to the end of the list (either from the OSTree commit, or from
the filesystem itself).