poll: Allow distributors to override the refspec to update from
Use "eos.checkpoint-target" as a metadata item on a commit with
the name of a new refspec to pull from next time the updater
is run with that commit deployed.

This allows for "checkpoint commits", eg, stop updating the OS
at refspec eos-history-v1 and set its "checkpoint target" to
eos-history-v2. When eos-history-v1 is deployed, the next time
we fetch we'll fetch from eos-history-v2 instead and then from
that point forward will continue to fetch from eos-history-v2
since that was the booted refspec.

(Modified by Philip Withnall to fix some minor issues found during code
review.)

https://phabricator.endlessm.com/T19437
https://phabricator.endlessm.com/T16682