test-common: Add helper functions to install from flatpak repositories
We're going to support installing flatpaks directly from the updater. In
order to test that, we'll need to be able to set up flatpak repositories
and install things from them, so this commit does exactly that. The functions
in flatpak-spawn.[ch] are wrappers around the flatpak command line utility
and the functions added to utils.[ch] are more general abstractions designed
to set up a flatpak repository with some pre-seeded apps, for use with testing.

https://phabricator.endlessm.com/T16682