installer: Add helper library to install already pulled flatpaks
When we auto-install flatpaks we will be doing so in a two-step process:
one step will merely pull the refs for the flatpaks while the user has
a network connection; the other will actually deploy all the refs
on reboot at the same time as a new system update is deployed. This
allows us to synchronize the removal of system components from the
OS image and ship them as flatpaks instead, completely transparently
to the user (since we're assuming that the flatpak deployments from
already pulled flatpaks will pretty much always succeed).

This helper library contains all the functions necessary to figure out
what flatpak actions need to be applied on reboot and perform them.

(Modified slightly by Philip Withnall to fix various small review
issues: see PR #129 for details.)

https://phabricator.endlessm.com/T16682