lib: Add flatpak helper functions to libeos-updater-util
This commit lands the first chunk of work to support auto-installation
of flatpaks on system updates. It introduces the concept of
a FlatpakRefAction, which is an action that can be done on a FlatpakRef
and is only applied once during the lifetime of the system. a FlatpakRefAction
can come from many different places in the system and might have different
priorities, causing it to clobber other FlatpakRefActions.

The idea is that on an update, the system will figure out all FlatpakRefActions
which are new and have not yet been applied, figure out the optimal
ordering, deduplicate the actions and then apply them all at once.

(Modified by Philip Withnall to fix some review issues: see PR #129.
Fixes to #includes, function naming, some leaks, some conformance to the
specification, and add some additional test cases.)

https://phabricator.endlessm.com/T16682