installer: Add eos-updater-flatpak-installer
This binary uses the functions in libeos-updater-flatpak-installer
to actually install flatpaks on reboot. It also has a "check" mode
which checks if the flatpak actions we claim to have applied are actually
fully-up-to-date on the system.

(Modified by Philip Withnall to improve formatting on man pages and fix
a few minor bugs found during review.)

https://phabricator.endlessm.com/T16682