fetch: Add a mechanism to atomically update flatpaks with OS upgrades
The updater now looks in
/var/share/eos-application-tools/flatpak-autoinstall.d/*,
/etc/eos-application-tools/flatpak-autoinstall.d/* and
/usr/share/eos-application-tools/flatpak-autoinstall.d/* for "auto-install"
files, which are a JSON formatted list of flatpak actions to apply upon an
update. It works by keeping a "state" file in
/var/lib/eos-application-tools/flatpak-autoinstall.progress which keeps
track of which actions have been applied so far.

See the included
eos-updater-flatpak-installer/eos-flatpak-autoinstall.schema.json for a
JSON Schema description of the autoinstall file.

Note that it assumes that the file is only ever append only - if you need
to remove a flatpak in a later update, you should add a new entry with
a "remove" action. It is completely valid to have a file that specifies
that a flatpak should be removed, then installed, then removed again -
the mechanism will squash down changes such that only a minimal amount
of work is done.

Also note that a file with the same name in a higher priority directory
will override all actions, including future actions included
in those files.

Before a system update is applied, the updater will now look inside
the commit to determine which flatpaks to install or remove. In the
case where flatpaks need to be installed, it will pull the relevant
ostree refs into the flatpak repository and only on success continue
with deployment of the system update. Then, once the deployment occurs,
a second script will run on boot which examines the contents of the
deployed flatpak autoinstall directories and will either deploy flatpaks
to remove them appropriately. Thus, an internet connection is only
required up to the point when the system update is deployed, but is not
required on the next reboot when flatpaks are actually installed or
removed.

(Modified by Philip Withnall to fix some minor issues with spec
compliance and the unit tests, plus some leak fixes.)

https://phabricator.endlessm.com/T16682