tests: Add tests for ensuring that flatpaks actually get updated
Prior tests were only checking that the updated ref was being pulled
into the repository, but were not exercising the relevant codepath
in the installer itself. These new tests actually run the installer
in its update mode and ensure that the newly updated flatpak
has contents that match what we expect.