util/flatpak: Move warnings out of unit tested code
Hoist the emission of warnings on skipping action entries up to a higher
level in the code. This means that the unit tests for the JSON parser
don’t have to deal with warnings, which makes it possible to test more
of the error handling paths in the parser and get almost full coverage
(see the following commits for improvements to the coverage to bring it
to 100%).

Do this by returning an additional array of the entries which were
skipped, so the caller can decide whether to warn about them or, in the
case of the unit tests, count them.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16682