src/fetch: Fix collection-ID-to-remote-name cache and checks
A few fixes to the
transition_pending_ref_action_collection_ids_to_remote_names() function.

The cache hash table was never actually being used, so do a lookup in it
so that it is.

Secondly, the function was erroring out if any lookup of a remote name
from a collection ID failed. Since we can’t (yet) guarantee that any
users have collection IDs specified in /ostree/repo/config, we can’t be
that strict with the check. Instead, print a message if the lookup
fails. Only throw an error if the lookup succeeds and its result doesn’t
match the remote name specified in the autoinstall JSON file.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20812