libeos-updater-util: Propagate failure to open the OSTree repository
Rather than aborting the updater with a g_error() (which is intended for
programming errors only), propagate the GError which indicates failure
to open /ostree/repo, and let the callers work out how to deal with it.
In the case of main(), this means returning a non-zero exit status and
bailing out of the updater (but slightly more gracefully than with
abort()).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20888