flatpak-updater: add eos-extra-settled.target as pre-requisite
eos-extra-settled.target is generated with a dependency on the SD
card being mounted at /var/endless-extra on "split" systems which
have their Flatpaks on the SD card. Adding this dependency makes
sure that the Flatpak system installation is ready/accessible
before we try and deploy the Flatpaks - although it has the side
effect that this unit won't complete on systems where the SD card
is missing.

https://phabricator.endlessm.com/T20696