src: Add support for download scheduling using mogwai-scheduled
Add a dependency on libmogwai-schedule-client, so that we can wait until
the download scheduler says it’s OK before we download OS updates.

Currently, pausing and resuming downloads based on direction from the
scheduler is not supported. Once a download is started, it’s proceeding
until it’s finished. This will change in future (not least once the
scheduler actually has support for changing its scheduling decisions on
the basis of new information.)

This does not change the behaviour of the existing
UpdateOnMobile=[true|false] configuration option for eos-autoupdater;
that will be updated later.

Signed-off-by: Philip Withnall <withnall@endlessm.com>
https://phabricator.endlessm.com/T20396