src: Fix declaration of handle_poll_volume() to avoid crashes
The signal type for handle-poll-volume is gdbus-codegen–generated, and
includes all the method call parameters as individual arguments. The
previous C declaration of handle_poll_volume() did not include these
arguments, meaning that the user_data argument received the wrong value
(the path passed to PollVolume()). That caused crashes.

Fix that by fixing the declaration to list all the right parameters.
This allows a bit of code simplification in surrounding code.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20396