autoupdater: Deprecate UpdateOnMobile= setting
Deprecate and remove support for this setting, since we now consult the
download scheduler about whether to download (and it factors in whether
we are on a mobile connection in its decision).

Keep the setting in the documentation for the configuration file (mark
it as deprecated), and aggressively remove it everywhere else.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20396