eos-autoupdater: Pass through the --force-update option to FetchFull
Now that the logic for forcing an update on a metered connection has
moved the updater (rather than the autoupdater), the autoupdater should
pass through its --force-update argument to the relevant place,
otherwise updates can never be forced.

In particular, this fixes the unit tests to ignore the scheduling
daemon. In future, we need to add more (scheduling specific) unit tests
which will use a mock scheduling daemon; but for the moment, the
existing tests should proceed as if always unmetered.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20396