src: Eliminate unnecessary EosMetadataFetchData class
It was used to pass around a few variables, but was used from multiple
threads, and was not thread safe. This moves us a step closer to having
a thread-safe poll function.

The need for this class was significantly reduced when the legacy
LAN/USB poll paths were removed from the codebase (commit 4b88c5667).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21123