src: Move GMainContext handling into a wrapper of poll_volume()
Make poll_volume_internal() be a more conventional function, which
doesn’t touch GMainContext or GTask, and just returns its return value
and a GError.

This makes the code a bit easier to follow and, crucially, fixes a bug
where the GMainContext it created internally would not be used by OSTree
as it wasn’t pushed as the thread default main context. The new wrapper
function pushes a single new thread default main context, which will be
used for everything in PollVolume().

This fixes a bug whereby the iteration loops using async_result_cb()
would never exit, as the iterating GMainContext never had an event to
handle.

The same treatment will be given to the Poll() code path in the
following few commits.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21123