eos-updater-ctl: replace lambdas with functions
If no user_data parameter is passed to .connect(), the function does not
receive that argument.

signal_emitted() is only called from this lambda, so we can just add the
extra argument and remove the lambda.

dump_daemon_properties() is called elsewhere, but the callback is
clearer written as a function anyway.

This also fixes the following warnings from flake8:

    src/eos-updater-ctl:96:66: E502 the backslash is redundant between brackets
    src/eos-updater-ctl:97:23: E127 continuation line over-indented for visual indent
    src/eos-updater-ctl:99:72: E502 the backslash is redundant between brackets
    src/eos-updater-ctl:100:23: E127 continuation line over-indented for visual indent