systemd: Don't use SystemCallFilter with the flatpak installer
Part of the design of the flatpak installer is that it was assumed
to never fail except in very exceptional cases, such disk space
running out. SystemCallFilter is a noble concept for reducing
attack surface area, but ultimately, we need the flatpak installer
to work correctly across updates, which means that changes in
the system calls that ostree or flatpak itself might use should
not break the flatpak installer due to security policy.

https://phabricator.endlessm.com/T21751