poll-common: Do not return error when no update has been found
Recent internal changes made it so that when no update exists, an error
is returned.
However, the users of that code expect no error to be set when no update
exists.

To fix that, this patch removes the setting of the error. So it returns
no update info nor error when no update is available.

https://phabricator.endlessm.com/T21857