eupv: Add autoinstall flatpaks to update USBs
When an OS update has flatpak(s) set to be updated or installed along
with it (via eos-updater-flatpak-installer), the flatpaks are required
for the update for succeed. So this commit adds the flatpaks to a USB
being created with eos-updater-prepare-volume so that USB updates work
even when the update in question has flatpaks to be autoinstalled. This
is accomplished using the GObject Introspection bindings of
libeos-updater-flatpak-installer, so that the logic that makes sense of
all the autoinstall files can stay there and not be duplicated.  Since
we can't make any assumptions about the OS version on the client
receiving the USB update, every flatpak configured to be automatically
updated or installed on OS update must be included.

https://phabricator.endlessm.com/T21756