flatpak: Ignore related refs that don't have a collection-id set
One of our invariants is that remote ref actions specified in
the eos-updater-flatpak-autoinstall.d(5) must have a collection-id
set. It is possible that we might need to fetch dependency refs from
other remotes and so we need to look at that remote's configuration
for the collection-id. If the remote does not have a collection-id
set then there is no way to continue without breaking the invariant,
so ignore the remote.