Cancel fetching when Mogwai asks it to not download
Fetching the updates was honoring the 'download-now' property from
Mogwai's Scheduled Entry for starting, but once it started, it'd not get
canceled even if Mogwai reported download-now=false for the entry.

This patch implements that cancellation by using a local cancellable
object for the fetching operation instead of the one that's passed to
the call, since the latter should only be cancelled by the Cancel()
D-Bus method (in order to avoid future issues).

When the fetch operation is canceled because of the entry's download-now
becoming false, the updater gets into error mode, with the error name
com.endlessm.Updater.Error.Cancelled.

https://phabricator.endlessm.com/T22237