eufi: Also delete autodelete extensions for refs marked uninstalled.
Instead of doing dependency resolution again like we do in the
fetch stage, Work out which extensions marked autodelete need
to be removed if a flatpak is uninstalled during the uninstallation
process. This removes a lot of the complexity and fragility around
ensuring that dependency resolution is correct and that the expected
metadata doesn't go out of sync between pulling the refs and rebooting.