test-common: Allow tests to specify the entire commit graph
Right now, test commits are created by specifying the most recent
commit number for a given ref in a hash table and then recursively
creating all the commits when the subserver is updated. This works
great for linear histories but is troublesome to deal with when we
want histories where we might introduce new branches with checkpoints
and create static deltas between them.

This commit introduces the concept of an EosTestUpdaterCommitGraph,
which contains a reverse adjacency list implemented with hash tables
(eg, every commit knows its parent, finding the children of a commit
is O(N)). A set of helper functions was introduced so that a
commit graph can be populated from the "leaf nodes" in the
refspec_to_commit hashtable, however the underlying representation
is now a graph. Also, the graph is walked explicitly breadth-first
from the root node down to create all the commits, as opposed to
implicitly and recursively creating commits for each of the parents.

All of this work allows tests to specify a non-linear history, which
will be important for the upcoming tests on refspec checkpoints
and continued commits on the non-checkpoint branch.