poll: Check booted branch first before following update refspec
Previously we would always unconditionally follow the upgrade refspec
if we were booted into a commit that had the update refspec as part
of its metadata. However, such an approach means that the client must
always be able to use the current updater on the checkpoint commit
to upgrade to the most recent commit. This leaves very little room
for error in case the updater on the current commit is not able
to process the new upgrade (for instance, incorrectly configured
flatpak remotes).

Considering that, we might want to "fix" the updater again before
we try upgrading on the branch that really requires it. That can
be done by adding new commits to the currently booted branch. However,
if we do that we need the updater to pay attention to the currently
booted branch and not the "upgrade" branch in the booted commit's
metadata.

This commit does exactly that. It wraps get_refspec_to_upgrade_on
with another function that first pulls on the currently booted branch
to see if there is a new commit and if so uses that, then, if there
is not a new commit, pulls the upgrade-refspec branch for new commits
and attempts to apply them.

https://phabricator.endlessm.com/T22295