poll-common: Operate on checkpoint refs, not refspecs
The eos.checkpoint-target commit metadata is created on the server side
while the remote is a property of the client's configuration. Following
a remote specified in the metadata would almost certainly be wrong and
provide the server with the ability to switch a client's remote.

Instead, only take the ref from the metadata like the
ostree.endoflife-rebase handling. If a full refspec with remote is
included, the remote is ignored and a warning is issued.

https://phabricator.endlessm.com/T19437