fetch: Ensure we can cancel checking the scheduler
The fetch operation couldn't be canceled because the check_scheduler was
blocked iterating its context (while waiting for Mogwai's signals).

These changes add a new GSource from the cancellable object to the
GMainContext, so when the cancellable is canceled the main context
iterates and thus allows for the operation to stop.

https://phabricator.endlessm.com/T21133