poll-common: Set upgrade ref correctly
ref_to_upgrade_on is a gchar, not an OstreeCollectionRef. Furthermore,
passing back the booted collection ref would not point to the checkpoint
ref. I guess this info is only used in the P2P path, which only looks at
the collection ref. Actually, it appears that UpdateRefInfo.ref isn't
actually used anywhere.

https://phabricator.endlessm.com/T19437