eos-updater-apply: Set remote branches option after deploy
This is mostly cosmetic since the only time the remote branches option
is used is when a pull is done with no refs specified. Still, people
expect it to match the ref in the origin refspec, and changing it to
match prevents inadvertently pulling an unused ref that would occupy
disk space. Errors from the option update are ignored since it has no
effect on the actual upgrade.

https://phabricator.endlessm.com/T22477