poll: Don't treat move to older commit on new branch as a downgrade
Previously we checked if the commit timestamp was always greater than
the booted commit timestamp before pulling it. This won't work in
the case that a commit is made on a new branch (1), then we make
another commit on the old branch (2) with a checkpoint pointing
to then new branch.

https://phabricator.endlessm.com/T22498